Camera ray

+++ nml_matrix/srcs/nml_mat_col_l2norm.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   nml_mat_col_l2norm.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nnakarac <nnakarac@student.42.fr>          +#+  +:+       +#+        */
+/*   By: nnakarac <nnakarac@42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/08 20:09:37 by nnakarac          #+#    #+#             */
-/*   Updated: 2023/04/30 18:47:17 by nnakarac         ###   ########.fr       */
+/*   Updated: 2023/05/05 00:51:31 by nnakarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,5 +30,5 @@ float	nml_mat_col_l2norm(t_nml_mat *m, unsigned int col)
 		sum += (m->data[i][col] * m->data[i][col]);
 		i++;
 	}
-	return (sqrt(sum));
+	return (sqrtf(sum));
 }
